+++ internal/app/loki-multi-tenant-proxy/auth.go
@@ -9,6 +9,7 @@ import (
 	"net/http"
 	"strings"
 
+	"github.com/coreos/go-oidc"
 	"github.com/giantswarm/loki-multi-tenant-proxy/internal/pkg"
 	"go.uber.org/zap"
 )
@@ -34,57 +35,68 @@ const (
 	// OrgIDKey Key used to pass loki tenant id though the middleware context
 	OrgIDKey key = iota
 	realm        = "Loki multi-tenant proxy"
+	readUser     = "read"
 )
 
 func Authentication(handler http.HandlerFunc, authConfig *pkg.Authn, logger *zap.Logger) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		token, ok := r.Header["X-Id-Token"]
+		var authenticationMode string
+		var user = "read" // default user if oauth mode
+		var pass = ""     // no password in oauth mode
 		if ok {
 			// OAuth token authentication mode (X-Id-Token header provided)
-			logger.Info("OAuth authentication type")
+			logger.Info("OAuth authentication mode")
 			logger.Info(fmt.Sprintf("Token = %s", token[0]))
-			// Decoding base64 jwt token
+			// Decode OAuth token payload section
 			payload, err := decodeOAuthToken(token[0])
 			if err != nil {
-				logger.Error("Error decoding token payload", zap.Error(err))
+				logger.Error(fmt.Sprintf("Error decoding token payload %s", token[0]), zap.Error(err))
 				return
 			}
 			// Token validation against Dex
-			err = validateOAuthToken(payload.Iss, payload.Aud)
+			err = validateOAuthToken(token[0], payload.Iss, payload.Aud, r.Context())
 			if err != nil {
-				logger.Error("Error while validating OAuth token against DEX", zap.Error(err))
+				logger.Error(fmt.Sprintf("Error while validating OAuth token against DEX %s", token[0]), zap.Error(err))
 				writeUnauthorisedResponse(w, "oauth")
 				return
 			}
-			//ctx := context.WithValue(r.Context(), OrgIDKey, orgID)
-			ctx := context.WithValue(r.Context(), OrgIDKey, "giantswarm")
-			handler(w, r.WithContext(ctx))
-			//OAuth(handler, authConfig, logger)
+			authenticationMode = "oauth"
+
 		} else {
-			// Default authentication mode => BasicAuth
-			logger.Info("BasicAuth authentication type")
-			user, pass, ok := r.BasicAuth()
-			authorized, orgID := isAuthorized(user, pass, authConfig)
-			if !ok || !authorized {
+			// BasicAuth authentication mode (X-Id-Token header not provided) - default mode (use for write path)
+			logger.Info("BasicAuth authentication mode")
+			user, pass, ok = r.BasicAuth()
+			if !ok {
 				writeUnauthorisedResponse(w, "basic")
 				return
 			}
-			ctx := context.WithValue(r.Context(), OrgIDKey, orgID)
-			handler(w, r.WithContext(ctx))
-			//BasicAuth(handler, authConfig, logger)
+			authenticationMode = "basic"
 		}
+		// Check if user is authorized to access Loki and retrieve OrgID
+		authorized, orgID := isAuthorized(user, pass, authConfig, authenticationMode)
+		if !authorized {
+			writeUnauthorisedResponse(w, authenticationMode)
+			return
+		}
+		ctx := context.WithValue(r.Context(), OrgIDKey, orgID)
+		handler(w, r.WithContext(ctx))
 	}
 }
 
-func isAuthorized(user string, pass string, authConfig *pkg.Authn) (bool, string) {
+// isAuthorized checks if the user is authorized to access Loki (BasicAuth mode)
+// and get OrgId to handle multi-tenant access
+func isAuthorized(user string, pass string, authConfig *pkg.Authn, authenticationMode string) (bool, string) {
 	for _, v := range authConfig.Users {
-		if subtle.ConstantTimeCompare([]byte(user), []byte(v.Username)) == 1 && subtle.ConstantTimeCompare([]byte(pass), []byte(v.Password)) == 1 {
+		// OAuth mode: retrieve user 'read' et get OrgID
+		// BasicAuth mode: check user and password passed in the request and get OrgID
+		if (authenticationMode == "oauth" && subtle.ConstantTimeCompare([]byte(user), []byte(v.Username)) == 1) ||
+			(authenticationMode == "basic" && subtle.ConstantTimeCompare([]byte(user), []byte(v.Username)) == 1 && subtle.ConstantTimeCompare([]byte(pass), []byte(v.Password)) == 1) {
 			if !authConfig.KeepOrgID {
 				return true, v.OrgID
 			} else {
 				return true, ""
 			}
-
 		}
 	}
 	return false, ""
@@ -98,16 +110,30 @@ func writeUnauthorisedResponse(w http.ResponseWriter, authenticationType string)
 	w.Write([]byte("Unauthorised\n"))
 }
 
+// decodeOAuthToken decodes the payload section of the OAuth token
 func decodeOAuthToken(token string) (Payload, error) {
 	// Get payload section from the token
 	payload := strings.Split(token, ".")[1]
-	payloadDecoded, _ := b64.URLEncoding.DecodeString(payload)
+	payloadDecoded, err := b64.RawURLEncoding.DecodeString(payload)
+	if err != nil {
+		return Payload{}, err
+	}
 
 	var p Payload
-	err := json.Unmarshal(payloadDecoded, &p)
+	err = json.Unmarshal(payloadDecoded, &p)
 	return p, err
 }
 
-func validateOAuthToken(dexUrl string, clientId string) error {
-	return nil
+// validateOAuthToken validates the OAuth token against Dex
+func validateOAuthToken(token string, dexUrl string, clientId string, ctx context.Context) error {
+	// Initialize a provider by specifying dex's issuer URL.
+	provider, err := oidc.NewProvider(ctx, dexUrl)
+	if err != nil {
+		return err
+	}
+	// Create an ID token parser, but only trust ID tokens issued to 'clientId'
+	idTokenVerifier := provider.Verifier(&oidc.Config{ClientID: clientId})
+	// Verify token validity
+	_, err = idTokenVerifier.Verify(ctx, token)
+	return err
 }