+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add OAuth token support for 'read' user.
+
 ## [0.2.0] - 2023-12-11
 
 ### Changed: